# frontend/src/lib/helper/typegenerator/api-docs.yaml
openapi: 3.0.1
info:
  title: Tutorial Management API
  description: This API exposes endpoints to be implemented in the frontend
  version: "3.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearer-key: []
...
paths:
...
  /update-user:
    put:
      tags:
        - person-controller
      summary: Endpoint for Admins to change/update a user
      description: Endpoint for Admins to change/update a user
      operationId: updateUser_1
      parameters:
        - name: personId
          in: query
          description: The ID of the User to update
          required: true
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          description: The new username
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: The new email
          required: false
          schema:
            type: string
        - name: permissions
          in: query
          description: The new Permissions
          required: false
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
              enum:
                - USER
                - GARDENER
                - ADMIN
        - name: password
          in: query
          description: The new Password
          required: false
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: A RESTResponse indicating Success
          content:
            "*/*":
              schema:
                type: string
    post:
      tags:
        - person-controller
      summary: Endpoint for Admins to change/update a user
      description: Endpoint for Admins to change/update a user
      operationId: updateUser
      parameters:
        - name: personId
          in: query
          description: The ID of the User to update
          required: true
          schema:
            type: string
            format: uuid
        - name: username
          in: query
          description: The new username
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: The new email
          required: false
          schema:
            type: string
        - name: permissions
          in: query
          description: The new Permissions
          required: false
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
              enum:
                - USER
                - GARDENER
                - ADMIN
        - name: password
          in: query
          description: The new Password
          required: false
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: A RESTResponse indicating Success
          content:
            "*/*":
              schema:
                type: string
...
  /update-settings:
    post:
      tags:
        - person-controller
      summary: Endpoint for a User to change his Account Settings.
      description: Endpoint for a User to change his Account Settings.
      operationId: updateSettings
      parameters:
        - name: person
          in: query
          description: Auto-injected Instance of the currently logged-in Person.
          required: true
          schema:
            $ref: "#/components/schemas/Person"
        - name: username
          in: query
          description: The new username
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: The new email
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The new Password
          required: false
          schema:
            type: string
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: A RESTResponse indicating Success
          content:
            "*/*":
              schema:
                type: string
...
  /create-user:
    post:
      tags:
        - person-controller
      summary: Create a new User
      description: Create a new User with the given credentials and permissions.
      operationId: create
      parameters:
        - name: username
          in: query
          description: The new Users username.
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The new Users password (needs to be already hashed).
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: The new Users email.
          required: true
          schema:
            type: string
        - name: permissions
          in: query
          description: The Permissions the new User should have.
          required: true
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
              enum:
                - USER
                - GARDENER
                - ADMIN
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: User successfully created.
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/CreatedUserResponse"
...
  /get-user-permissions:
    get:
      tags:
        - person-controller
      summary: Get all permissions of a user
      operationId: getUserPermissions
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Returns all permissions of a user
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PermissionResponse"
...
  /get-all-users:
    get:
      tags:
        - person-controller
      summary: Get all users
      description:
        "Endpoint for Admins to get all users, except the calling user\
        \ himself."
      operationId: getAllUsers
      parameters:
        - name: person
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Person"
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully retrieved all users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
...
  /get-all-permissions:
    get:
      tags:
        - person-controller
      summary: Get all permissions
      description: |-
        Endpoint for Admins to get all possible Permission so that they don't
         need to be changed manually on frontend.
      operationId: getAllPermissions
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully retrieved all permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
  /get-all-gardener:
    get:
      tags:
        - person-controller
      summary: Get all gardeners
      description: Endpoint for Admins to get all gardeners
      operationId: getAllGardener
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Successfully retrieved all gardeners
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
...
  /delete-user:
    delete:
      tags:
        - person-controller
      summary: Endpoint for Admins to delete a user.
      description: Endpoint for Admins to delete a user.
      operationId: deleteUser
      parameters:
        - name: personId
          in: query
          description: The ID of the User to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "500":
          description: Internal Server Error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: |-
            A RestResponse indicating whether the operation was
                 successful or not.
          content:
            "*/*":
              schema:
                type: string
...
components:
  schemas:
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    Person:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          writeOnly: true
        token:
          type: string
          format: uuid
          readOnly: true
        permissions:
          uniqueItems: true
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
        email:
          type: string
        sensorStationPersonReferences:
          type: array
          items:
            $ref: "#/components/schemas/SensorStationPersonReference"
        personId:
          type: string
          format: uuid
    SensorStationPersonReference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inDashboard:
          type: boolean
        assigned:
          type: boolean
        deleted:
          type: boolean
...
    CreatedUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        token:
          type: string
          format: uuid
        permissions:
          uniqueItems: true
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
      description: ""
...
    PermissionResponse:
      type: object
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/GrantedAuthority"
      description: ""
...
    ListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
      description: |-
        A Response Container for sending a List of items fetched on the server side.
         Only if the success-Field is set to true then a List is sent.
...
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT